test/requirements.txt
pytest>=8.0
pytest-timeout>=2.3
requests>=2.31
openpyxl>=3.1
lxml>=5.0
psycopg2-binary>=2.9
